Use stdlib sin and cos. Convert double to float.
Changing double to float only for readings.

+++ include/csm/laser_data.h
@@ -28,7 +28,7 @@ struct laser_data {
 	double * restrict theta;
 	
 	int    * restrict valid;
-	double * restrict readings;
+	float  * restrict readings;
 	
 	int    * restrict cluster;
